+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 * Throw ContextualRuntimeException in user function visitor 
 
+### Changed
+
+* Add support for escaped variable names
+
 ## 0.1.10-2 - 2018-09-04
 
 ### Changed